Development

+++ src/main/java/com/seriesrecap/SeriesRecap/Repositories/TypesRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TypesRepository extends JpaRepository <Types, Long> {
+public interface TypesRepository extends JpaRepository <Types, Integer> {
 }
